pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "memeplotlib"
version = "0.1.0"
description = "Memes with Python's matplotlib"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Brian Keegan" },
]
keywords = ["meme", "matplotlib", "memegen", "image-macro"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics",
    "Framework :: Matplotlib",
]
dependencies = [
    "matplotlib>=3.0.0",
    "requests>=2.20.0",
    "numpy>=1.16.0",
    "platformdirs>=2.0.0",
    "Pillow>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.0",
    "responses>=0.20.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/brianckeegan/memeplotlib"
Repository = "https://github.com/brianckeegan/memeplotlib"
Issues = "https://github.com/brianckeegan/memeplotlib/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/memeplotlib"]

[tool.hatch.build.targets.wheel.shared-data]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (require network)",
    "integration: marks integration tests",
]

[tool.ruff]
target-version = "py310"
line-length = 99
